+++ product/src/domain/product-sub-category-domain.ts
@@ -6,36 +6,64 @@ import { ProductSubCategoryDatabaseLayer } from '../database-layer/product-sub-c
 export class ProductSubCategoryDomain {
 
     static async createProductSubCategory(req: Request, res: Response) {
-        const ProductSubCategory = await ProductSubCategoryDatabaseLayer.createProductSubCategory(req);
-        res.status(201).send(ProductSubCategory);
+        console.log('1');
+        
+        const permission = await ProductSubCategoryDatabaseLayer.categoryCheck(req);
+        if (permission.isCreate == true || permission.isSuperAdmin == true) {
+            const ProductSubCategory = await ProductSubCategoryDatabaseLayer.createProductSubCategory(req);
+            res.status(201).send(ProductSubCategory);
+        } else {
+            throw new BadRequestError('You don\'t have rights to create category');
+        }
     }
 
     static async updateProductSubCategory(req: Request, res: Response) {
         if (!mongoose.isValidObjectId(req.params.id)) {
             throw new BadRequestError('Requested id is not id type');
         }
-        const data = await ProductSubCategoryDatabaseLayer.updateProductSubCategory(req,req.params.id);
-        res.status(201).send(data);
+        const permission = await ProductSubCategoryDatabaseLayer.categoryCheck(req);
+        if (permission.isUpdate == true || permission.isSuperAdmin == true) {
+            const data = await ProductSubCategoryDatabaseLayer.updateProductSubCategory(req, req.params.id);
+            res.status(201).send({ updated: true });
+        } else {
+            throw new BadRequestError('You don\'t have rights to create category');
+        }
     }
 
     static async deleteProductSubCategory(req: Request, res: Response) {
         if (!mongoose.isValidObjectId(req.params.id)) {
             throw new BadRequestError('Requested id is not id type');
         }
-        await ProductSubCategoryDatabaseLayer.deleteProductSubCategory(req.params.id);
-        res.status(201).send({ deleted: true });
+        const permission = await ProductSubCategoryDatabaseLayer.categoryCheck(req);
+        if (permission.isDelete == true || permission.isSuperAdmin == true) {
+            await ProductSubCategoryDatabaseLayer.deleteProductSubCategory(req.params.id);
+            res.status(201).send({ deleted: true });
+        } else {
+            throw new BadRequestError('You don\'t have rights to create category');
+        }
     }
 
     static async getProductSubCategoryList(req: Request, res: Response) {
-        const ProductSubCategory =  await ProductSubCategoryDatabaseLayer.getProductSubCategoryList(req);
+        const ProductSubCategory = await ProductSubCategoryDatabaseLayer.getProductSubCategoryList(req);
         res.status(201).send(ProductSubCategory);
     }
     
+    static async getProductSubCategoryId(req: Request, res: Response) {
+        if (!mongoose.isValidObjectId(req.params.id)) {
+            throw new BadRequestError('Requested id is not id type');
+        }
+        const ProductSubCategory = await ProductSubCategoryDatabaseLayer.getProductSubCategoryId(req,req.params.id);
+        res.status(201).send(ProductSubCategory);
+    }
     static async getProductCategoryIdList(req: Request, res: Response) {
         if (!mongoose.isValidObjectId(req.params.id)) {
             throw new BadRequestError('Requested id is not id type');
         }
-        const ProductSubCategory =  await ProductSubCategoryDatabaseLayer.getProductCategoryIdList(req,req.params.id);
+        const ProductSubCategory = await ProductSubCategoryDatabaseLayer.getProductCategoryIdList(req, req.params.id);
+        res.status(201).send(ProductSubCategory);
+    }
+    static async getProductSubCategoryActiveList(req: Request, res: Response) {
+        const ProductSubCategory = await ProductSubCategoryDatabaseLayer.getProductSubCategoryActiveList(req);
         res.status(201).send(ProductSubCategory);
     }
